pyproject.toml
[project]
name = "bootdev-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google>=3.0.0",
    "google-genai>=1.67.0",
    "google-generativeai>=0.8.6",
    "nltk==3.9.1",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
]
